+++ src/meico/audio/lame/mp3/GenreListHandler.java
@@ -1,5 +1,0 @@
-package meico.audio.lame.mp3;
-
-public interface GenreListHandler {
-  void genre_list_handler(int num, String name);
-}